requirements.txt
pillow
transformers
torch
pywaggle[vision]
